+++ components/product/LikeButton/useLike.ts
@@ -1,18 +1,16 @@
-import { Tooltip } from "@mui/material";
-import Router from "next/router";
-import { IoMdHeart, IoMdHeartEmpty } from "react-icons/io";
-import styles from "@styles/ProductDetail.module.css";
 import { productAPI } from "api/product";
 import { useToken } from "hooks/useToken";
+import Router from "next/router";
 
-interface Props {
+export const useLike = ({
+  isLike,
+  productId,
+}: {
   isLike?: boolean;
-  id: string;
-}
-
-const LikeButton = ({ isLike, id }: Props) => {
+  productId: string;
+}) => {
   const token = useToken();
-  const onClickLikeButton = async () => {
+  const onClick = async () => {
     if (token) {
       let wish = true;
       let alertMessage = "❤️ 위시리스트에 추가되었습니다.";
@@ -22,14 +20,14 @@ const LikeButton = ({ isLike, id }: Props) => {
       }
       try {
         const { data } = await productAPI.updateWish({
-          productId: id,
+          productId,
           wish,
         });
         const { success } = data;
         if (success) {
           Router.replace(
-            `/products/${id}?alert=${alertMessage}`,
-            `/products/${id}`
+            `/products/${productId}?alert=${alertMessage}`,
+            `/products/${productId}`
           );
         } else {
           alert("⚠️ 위시리스트 추가에 실패했습니다. 다시 시도해주세요.");
@@ -41,17 +39,5 @@ const LikeButton = ({ isLike, id }: Props) => {
       alert("🔒 로그인이 필요합니다.");
     }
   };
-  return (
-    <Tooltip title="위시리스트" arrow>
-      <button onClick={onClickLikeButton}>
-        {isLike ? (
-          <IoMdHeart className={styles.heartIcon} />
-        ) : (
-          <IoMdHeartEmpty className={styles.heartIcon} />
-        )}
-      </button>
-    </Tooltip>
-  );
+  return onClick;
 };
-
-export default LikeButton;
